+++ src/collective/collectionfilter/tiles/maps.py
@@ -3,12 +3,9 @@ from Products.CMFPlone.resources import add_bundle_on_request
 from Products.CMFPlone.utils import get_top_request
 from collective.collectionfilter.baseviews import BaseMapsView
 from collective.collectionfilter.interfaces import ICollectionMapsSchema
-from collective.collectionfilter.tiles import DictDataWrapper
+from collective.collectionfilter.tiles import BaseFilterTile
 from plone.supermodel.model import Schema
-from plone.tiles.tile import PersistentTile
 from zope.interface import implementer
-
-import plone.api
 
 
 class IMapsTile(Schema, ICollectionMapsSchema):
@@ -16,7 +13,7 @@ class IMapsTile(Schema, ICollectionMapsSchema):
 
 
 @implementer(IMapsTile)
-class MapsTile(PersistentTile, BaseMapsView):
+class MapsTile(BaseFilterTile, BaseMapsView):
 
     def __init__(self, context, request):
         super(MapsTile, self).__init__(context, request)
@@ -24,24 +21,3 @@ class MapsTile(PersistentTile, BaseMapsView):
         #       called before this one; subrequest problem
         top_request = get_top_request(request)
         add_bundle_on_request(top_request, 'bundle-leaflet')
-
-    @property
-    def edit_url(self):
-        if not plone.api.user.has_permission(
-            'cmf.ModifyPortalContent',
-            obj=self.context
-        ):
-            return None
-        return self.url.replace('@@', '@@edit-tile/')
-
-    @property
-    def settings(self):
-        return DictDataWrapper(self.data)
-
-    @property
-    def filter_id(self):
-        return self.id
-
-    @property
-    def reload_url(self):
-        return self.url